ViewModel分割とSettingsの整形
